+++ libraries/ml/src/MCL/mcl-editor.ts
@@ -4,7 +4,7 @@ import {DimReductionEditorOptions} from '../functionEditors/dimensionality-reduc
 import {MultiColumnDimReductionEditor}
   from '../multi-column-dimensionality-reduction/multi-column-dim-reduction-editor';
 import {getGPUAdapterDescription} from '@datagrok-libraries/math/src/webGPU/getGPUDevice';
-import {defaultMCLOptions} from './marcov-cluster';
+import {defaultMCLOptions} from './markov-cluster';
 
 
 export class MCLEditor extends MultiColumnDimReductionEditor {