+++ packages/Charts/src/viewers/sunburst/sunburst-viewer.ts
@@ -18,6 +18,7 @@ type onClickOptions = 'Select' | 'Filter';
 })
 
 export class SunburstViewer extends EChartViewer {
+  private renderQueue: Promise<void> = Promise.resolve();
   hierarchyColumnNames: string[];
   hierarchyLevel: number;
   onClick: onClickOptions;
@@ -29,7 +30,7 @@ export class SunburstViewer extends EChartViewer {
     this.initCommonProperties();
     this.initEventListeners();
 
-    this.hierarchyColumnNames = this.addProperty('hierarchyColumnNames', DG.TYPE.COLUMN_LIST);
+    this.hierarchyColumnNames = this.addProperty('hierarchyColumnNames', DG.TYPE.COLUMN_LIST, null, {columnTypeFilter: DG.TYPE.CATEGORICAL});
     this.hierarchyLevel = 3;
     this.onClick = <onClickOptions> this.string('onClick', 'Select', { choices: ['Select', 'Filter'] });
     this.inheritFromGrid = this.bool('inheritFromGrid', true, {category: 'Color'});
@@ -40,6 +41,9 @@ export class SunburstViewer extends EChartViewer {
         {
           type: 'sunburst',
           nodeClick: false,
+          emphasis: {
+            focus: 'series',
+          },
           label: {
             rotate: 'radial',
             fontSize: 10,
@@ -314,17 +318,24 @@ export class SunburstViewer extends EChartViewer {
     return this.dataFrame.columns.length >= 1;
   }
 
-  render() {
+  render(): void {
+    this.renderQueue = this.renderQueue
+      .then(() => this._render());
+  }
+
+  private async _render() {
     if (this.hierarchyColumnNames?.some((colName) => !this.dataFrame.columns.names().includes(colName)))
       this.hierarchyColumnNames = this.hierarchyColumnNames.filter((value) => this.dataFrame.columns.names().includes(value));
+
     if (this.hierarchyColumnNames == null || this.hierarchyColumnNames.length === 0)
       return;
 
-    this.handleStructures(this.getSeriesData()).then((data) => {
-      this.option.series[0].data = data;
-      this.option.series[0].label.formatter = (params: any) => this.formatLabel(params);
-      this.chart.setOption(this.option);
-    });
+    const seriesData = await this.getSeriesData();
+    const data = await this.handleStructures(seriesData);
+
+    this.option.series[0].data = data;
+    this.option.series[0].label.formatter = (params: any) => this.formatLabel(params);
+    this.chart.setOption(this.option);
   }
 
   detach() {