+++ packages/UITests/src/viewers/viewers.ts
@@ -29,6 +29,15 @@ category('Viewers: Core Viewers', () => {
   }
 });
 
+category('Viewers', ()=> {
+  test('Viewers issues', async ()=> {
+    await DG.Test.testViewerProperties(grok.data.demo.demog(100), DG.VIEWER.BOX_PLOT, {valueColumnName: 'started'});
+    await DG.Test.testViewerProperties(grok.data.demo.demog(100), DG.VIEWER.GRID, {sortByColumnNames: ['age']});
+    await DG.Test.testViewerProperties(grok.data.demo.demog(100), DG.VIEWER.GRID, {pinnedRowColumnNames: ['subj']});
+    await DG.Test.testViewerProperties(grok.data.demo.demog(100), DG.VIEWER.PC_PLOT, {colorColumnName: 'started'});
+  });
+});
+
 category('Viewers', () => {
   let df: DG.DataFrame;
   let tv: DG.TableView;