+++ libraries/utils/CHANGELOG.md
@@ -1,5 +1,25 @@
 # utils changelog
 
+## 4.2.27 (2024-08-16)
+
+Invocation time refactoring
+
+## 4.2.26 (2024-08-15)
+
+Invocation time refactoring
+
+## 4.2.25 (2024-08-15)
+
+Added invocation time to tests 
+
+## 4.2.24 (2024-08-09)
+
+Forms viewer: Don't use color coding with molecule columns
+
+## 4.2.23 (2024-08-09)
+
+Forms viewer: Ability to resize header, sending event on input click
+
 ## 4.2.22 (2024-08-09)
 
 Added stress test invocation method