+++ packages/EDA/CHANGELOG.md
@@ -6,6 +6,14 @@
 * Reduced bias in the PLS regression
 * Fixed linear regression fails
 
+## 1.1.33 (2024-08-14)
+
+Bump ml version to 6.6.21
+
+## 1.1.32 (2024-08-13)
+
+Add XGBoost to predictive modeling tools
+
 ## 1.1.31 (2024-08-06)
 
 Add MCL layout with force directed graph and atlas