+++ packages/Peptides/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Peptides changelog
 
+## 1.17.29 (2024-08-14)
+
+Update MCL implementation
+
 ## 1.17.28 (2024-08-08)
 
 * Downgrade API version